Fix compilation of a22abb2-ffi on non Windows platforms for CI
We need to use a specific Windows API to allocate a string in
a22abb2-ffi and send it through the FFI to the C# side.

This breaks the Travis CI which does not run on Windows and
therefore cannot compile a22abb2-ffi.

To fix this, I added a stub allocation function that panics on
non-Windows platform but still compiles. This is fine because this
code is used by the Windows application and will not be run by the
CI.