simplify: Fix broken simplify for exponents
Previously, we would simplify `x^0` into `1` for every x. The problem is
that x can be 0 and in that case the simplification is invalid because `0^0`
is not possible but `1` is possible. I made it so that the simplification only
enables when we know for sure that x is not 0.

Also, I added another simplification that I forgot to do earlier: `x^1 = x`.