misc: Handle unknown constants in the `eval` binary
Previously, the `eval` binary would panic with an expression such as
`sin(x)` because it could not evaluate the expression because of the
unknown constant `x`.

Now, instead of panicking, we continue to show the result of the
simplification process but if the evaluation fails, we just print
the error.