simplify: Fix simplification of `sqrt((pi + e)^2)` not working
Now that I added expansion of exponents, the simplification of the
expression in the commit title would no longer work because it would
first simplify the inner expression into `pi^2 + pi*e + e*pi + e^2`
and then would not be able to simplify the square root of that back
into the original inner expression.

I solved this by checking for the "sub exponent" before doing the
expansion of the left term of an exponentiation during simplification.

Note that I've refactored the code into a long function instead of using
Rust's `match` like before because it makes the order of operations
clearer.