simplify: More fixes to preserve term order in sum
When we expand a product, we were doing the cartesian product in the
wrong order: `(a+b)(c+d)` must be simplified into `ac+ad+bc+bd`, not
`ca+cb+da+db`.

Only swap the order of a product when it can benefit us for the
simplification step: I changed the `>` into a `>=` in line 286 of
simplify.rs.