display: Fix check for implicit multiplication
I forgot to reset the `previous_was_int` variable when the node
is not an integer, which means that all multiplications after an
implicit multiplication were implicit too.

Now, it works correctly: it resets the flag if the node is not an
integer so that the next multiplication is an explicit multiplication.

For `2pi*e`:
- before this commit: `2pie`
- now: `2pi * e`