simplify: Put number at beginning in sum/product
I like `2pi` better than `pi * 2`, so I made it so that when it folds
numbers, it returns the former AST instead of the second.

Note that instead of always putting the numbers at the beginning, we
should probably have a smart algorithm that puts the final folded number
at position of the first number or something like that.