simplify: Add a flag to know whether the simplification step did anything
The simplification can be successful but we want to know if it really did
anything so that we can show the new simplified expression to the user.

I've had to refactor a lot of things to make this work. Now, `simplify`
returns Ãa `SimplifySuccess` on success instead of a `Node`.This
`SimplifySuccess` contains a `did_something` flag which tells us if the
function really did anything useful.