ui: Remove sub-optimal heuristic to determine whether to show approximation or not
Before, we checked if the approximation string was equal to the
simplified expression string or to the input expression string,
and if it was the case, we would not show the approximation
because it would be useless.

To me, comparing expression in strings feels error prone, so
instead, I replaced it by checking, in the Rust side, if the
simplified expression is an integer, in which case there is no
need to show an approximation because it would for sure be the
same as the simplified expression.