Run the calculations on a thread pool
Previously, we would run the calculations on the UI thread which
meant that big calculations would block the UI from responding.
Now, we run the calculations on a background thread on the thread
pool, using .NET's `Task.Run` method.

In order to prevent race conditions, each time we want to run a
new task, we cancel the previous one that was running.